requirements.txt
Dependencias
altgraph==0.17.5
distlib==0.4.0
et_xmlfile==2.0.0
filelock==3.20.3
openpyxl==3.1.5
packaging==26.0
pefile==2024.8.26
platformdirs==4.5.1
psycopg2==2.9.11
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
PyQt6==6.10.1
PyQt6-Qt6==6.10.1
PyQt6_sip==13.10.3
python-dotenv==1.2.1
pywin32-ctypes==0.2.3
setuptools==80.10.1
virtualenv==20.36.1
